Create files for each theme function
