+++ src/theme/index.js
@@ -1,0 +1,4 @@
+export { default as convert } from './convert.js'
+export { default as getThemeFromName } from './getThemeFromName.js'
+export { default as getAvailableThemes } from './getAvailableThemes.js'
+export { default as logArray } from './logArray.js'
